Final Commit
I think this will be the final commit for this code. I think this code was able to demonstrating its ability to predict the price of rental property in Vancouver. At this moment there are some sub objectives I would like to do, but I think I want to focus on some other aspects of machine learning for the time being.